chore: external layer refactor
